fix something that's not 64bit-safe

+++ src/pygame_sdl2/display.pyx
@@ -335,7 +335,7 @@ cdef class Window:
         """
 
         if not always:
-            if <unsigned long> SDL_GL_GetCurrentContext():
+            if <intptr_t> SDL_GL_GetCurrentContext():
                 return False
 
         self.gl_context = SDL_GL_CreateContext(self.window)
